line_app/requirements.txt
paho-mqtt
uvicorn
python-dotenv

# for linebot
flask
line-bot-sdk==3.13.0

pillow
# flask-socketio

pymongo

flask-pymongo
flask-login
flask-wtf
# Werkzeug==2.1.1

aift
